littlelemon/restaurant/urls.py
from django.urls import path
from rest_framework.authtoken.views import obtain_auth_token
from . import views


urlpatterns = [
    path('', views.index, name='index'),
    path('menu/', views.MenuItemView.as_view(), name='menu'),
    path('menu/<int:pk>/', views.SingleItemView.as_view(), name='single-item'),
    path('api-token-auth/', obtain_auth_token),
]
